+++ src/url.js
@@ -1,4 +1,4 @@
-let template  = require('./template')
+let parameterizeRoute  = require('./parameterizeRoute')
 let trimRight = /\/$/
 let trimLeft  = /^\//
 
@@ -6,5 +6,5 @@ module.exports = function (base, path, params) {
   base = base.replace(trimRight, '')
   path = path.replace(trimLeft, '')
 
-  return template(base + '/' + path, params).replace(trimRight, '')
+  return parameterizeRoute(base + '/' + path, params).replace(trimRight, '')
 }